requirements.txt
Django==2.2.17
django-cors-headers==3.5.0
django-environ==0.4.5
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2020.4
sqlparse==0.4.1
whitenoise==5.2.0
